+++ toolchain/shader_compiler/include/shader_compiler/assert.h
@@ -1,3 +1,5 @@
 #pragma once
 
 #include <common/assert.h>
+
+#define _assert(assertion) assert(DESCRIPTION(assertion), #assertion)
