+++ toolchain/shader_compiler/include/shader_compiler/compiler/syntax.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <shader_compiler/defines.h>
+#include <shader_compiler/compiler/compiler_ctx.h>
+#include <phymac_parser/v3d_generic.h>
+
+SC_API void syntax(v3d_generic_node_t* node, compiler_ctx_t* ctx);
