+++ toolchain/shader_compiler/include/shader_compiler/utilities/misc.h
@@ -1,0 +1,6 @@
+
+#pragma once
+
+#include <shader_compiler/defines.h>
+
+static inline u32 min(u32 v1, u32 v2) { return (v1 > v2) ? v2 : v1; }
