+++ toolchain/shader_compiler/shaders/depth_reflection.v3dshader
@@ -27,6 +27,7 @@ Shader
 	{
 		SubPass
 		{
+			[NoParse]
 			GraphicsPipeline
 			{
 				colorBlend
