+++ toolchain/shader_compiler/shaders/lit_color.v3dshader
@@ -24,6 +24,7 @@ Shader
 	{
 		SubPass
 		{
+			[NoParse]
 			GraphicsPipeline
 			{
 				colorBlend
