+++ toolchain/shader_compiler/shaders/point_depth_reflection.v3dshader
@@ -27,6 +27,7 @@ Shader
 	{
 		SubPass
 		{
+			[NoParse]
 			GraphicsPipeline
 			{
 				colorBlend
@@ -41,6 +42,7 @@ Shader
 				}
 			}
 
+			[NoParse]
 			glsl
 			{
 				#stage vertex
