+++ toolchain/shader_compiler/shaders/skybox.v3dshader
@@ -21,6 +21,7 @@ Shader
 	{
 		SubPass
 		{
+			[NoParse]
 			GraphicsPipeline
 			{
 				colorBlend
